Initial JS implementation.
The Builder seems a bit lrage, and makes a lot of Sets that will
probably be unused. The other option is to make them null.

Likewise, the MdnSearchResponse is very very thin. I'm not quite
sure it is useful.

It all also has to be documented still. But there's absolutely
nothing complicated here, so the documentation should by mostly
rote.